WIP navigaiton starting to work
